+++ src/compiler/parser/mod.rs
@@ -12,9 +12,15 @@ mod lines_2000_3000;
 pub use lines_2000_3000::*;
 mod lines_3000_4000;
 pub use lines_3000_4000::*;
-mod lines_4000_6000;
-pub use lines_4000_6000::*;
-mod lines_6000_8000;
-pub use lines_6000_8000::*;
-mod lines_8000_end;
-pub use lines_8000_end::*;
+mod lines_4000_5000;
+pub use lines_4000_5000::*;
+mod lines_5000_6000;
+pub use lines_5000_6000::*;
+mod lines_6000_7000;
+pub use lines_6000_7000::*;
+mod lines_7000_8000;
+pub use lines_7000_8000::*;
+mod lines_8000_9000;
+pub use lines_8000_9000::*;
+mod lines_9000_end;
+pub use lines_9000_end::*;
